+++ src/outputRendering/OutputRenderingManager.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { SimpleOutputRenderer } from './OutputRenderer';
+import SimpleOutputRenderer from './SimpleOutputRenderer';
 
 export default class OutputRenderingManager {
   constructor({ settings }) {
